requirements.txt
# Core Dependencies (Python 3.13 compatible)
numpy>=1.26.0
pandas>=2.1.0
python-dotenv>=1.0.0

# Geospatial Libraries
rasterio>=1.3.9
geopandas>=0.14.0
shapely>=2.0.2
pyproj>=3.6.1
sentinelsat>=1.2.1
sentinelhub>=3.10.0
folium>=0.15.0

# Machine Learning (Python 3.13 compatible)
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.35.0
scikit-learn>=1.3.2
opencv-python>=4.8.1.78

# Image Processing
Pillow>=10.1.0
scikit-image>=0.22.0

# API and Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
aiofiles>=23.2.1
httpx>=0.25.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
sqlalchemy>=2.0.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Report Generation
reportlab>=4.0.7
fpdf2>=2.7.6
jinja2>=3.1.2

# Utilities
tqdm>=4.66.1
requests>=2.31.0
click>=8.1.7
pyyaml>=6.0.1
